all mistakes are corrected
