+++ adminUserSite/src/main/java/com/example/adminusersite/servlets/crud/CreateUserByAdminServlet.java
@@ -2,23 +2,31 @@ package com.example.adminusersite.servlets.crud;
 
 import com.example.adminusersite.dao.UserDao;
 import com.example.adminusersite.dao.UserDaoImpl;
+import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.SneakyThrows;
 
+import java.io.IOException;
 
 @WebServlet("/create")
 public class CreateUserByAdminServlet extends HttpServlet {
     @Override
-    @SneakyThrows
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
-        String uname = req.getParameter("login");
-        String upass = req.getParameter("password");
         UserDao userDao = new UserDaoImpl();
 
-        userDao.createUser(uname, upass);
-        req.getRequestDispatcher("sign-in").forward(req, resp);
+        userDao.createUser(req.getParameter("login"), req.getParameter("password"));
+
+        try {
+            req.getRequestDispatcher("sign-in").forward(req, resp);
+        } catch (ServletException | IOException e) {
+            try {
+                req.setAttribute("message", "something went wrong");
+                req.getRequestDispatcher("errorPage.jsp").forward(req, resp);
+            } catch (ServletException | IOException ex) {
+                throw new RuntimeException(ex);
+            }
+        }
     }
 }
