+++ adminUserSite/src/main/java/com/example/adminusersite/servlets/crud/CreateUserServlet.java
@@ -3,29 +3,35 @@ package com.example.adminusersite.servlets.crud;
 import com.example.adminusersite.dao.UserDao;
 import com.example.adminusersite.dao.UserDaoImpl;
 import com.example.adminusersite.user.User;
+import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.SneakyThrows;
+
+import java.io.IOException;
 
 @WebServlet("/register")
 public class CreateUserServlet extends HttpServlet {
     @Override
-    @SneakyThrows
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
-        String uname = req.getParameter("login");
-        String upass = req.getParameter("password");
+        String username = req.getParameter("login");
+        String password = req.getParameter("password");
         UserDao userDao = new UserDaoImpl();
 
-        User user = userDao.findUser(uname, upass);
+        User user = userDao.findUser(username, password);
 
-        if (uname.equals(user.getU_name())) {
+        if (username.equals(user.getU_name())) {
             req.setAttribute("message", "this user already exist");
         } else {
-            userDao.createUser(uname, upass);
+            userDao.createUser(username, password);
             req.setAttribute("message", "user is created");
         }
-        req.getRequestDispatcher("errorPage.jsp").forward(req, resp);
+
+        try {
+            req.getRequestDispatcher("errorPage.jsp").forward(req, resp);
+        } catch (ServletException | IOException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 }
