+++ adminUserSite/src/main/webapp/errorPage.jsp
@@ -1,7 +1,7 @@
 <%@ page contentType="text/html;charset=UTF-8" %>
 <html>
 <head>
-    <title>Title</title>
+    <title>Message Page</title>
     <link rel="stylesheet" href="index.css">
 </head>
 <body>
