+++ adminUserSite/src/main/webapp/register.jsp
@@ -1,7 +1,7 @@
 <%@ page contentType="text/html;charset=UTF-8" %>
 <html>
 <head>
-    <title>Title</title>
+    <title>Registration</title>
     <link rel="stylesheet" href="index.css">
 </head>
 <body>
