+++ adminUserSite/src/main/webapp/userPage.jsp
@@ -2,7 +2,7 @@
 <%@ page contentType="text/html;charset=UTF-8" %>
 <html>
 <head>
-    <title>Title</title>
+    <title>User Page</title>
     <link rel="stylesheet" href="index.css">
 </head>
 <body>
